% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_tab.R
\name{cross_tab}
\alias{cross_tab}
\title{Make Cross Tabulations, Possibly with Weights, and Return as Tidy Data Frame}
\usage{
cross_tab(data, out, preds = NULL, wt = NULL, form = c("t", "p"))
}
\arguments{
\item{data}{A data.frame or tibble.}

\item{out}{Outcome variable name, as string.}

\item{preds}{Vector of predictor variable names, as string.}

\item{wt}{Weight variable name, as string.}

\item{form}{Format of resulting tibble. t for tidy (data analysis) or p for 
pretty (presentation).}
}
\value{
A tibble of counts and proportions of out, perhaps at numerous levels
or combinations of levels of preds.
}
\description{
Behaves very similarly to the descr::crosstab function, but it does not round
weighted group sizes to the nearest whole number.  Other CrossTab functions 
did not return values in a tidy data frame, so I adapted a dplyr/tidyr pipe 
chain to do so.
}
\examples{
set.seed(1839)
n <- 500
w <- rgamma(n, 1)
data <- data.frame(
  y = sample(letters[1:2], n, TRUE),
  x1 = sample(letters[5:8], n, TRUE),
  x2 = sample(letters[9:12], n, TRUE),
  w = w + (1 - mean(w))
)

cross_tab(data, "y", c("x1", "x2"), "w", "p")
}
