% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Get Rake Weights, Given Data and Target Population Margins}
\usage{
get_weights(data, targets, na_warn = TRUE)
}
\arguments{
\item{data}{Individual-level data.frame. Each row is unit of observation,
each column is a variable. The names of the variables in targets must be
in the data verbatim. Note that NAs are treated with listwise deletion.}

\item{targets}{Data.frame, containing three columns. The first column
must be the names of the variables on which one wants to weight. The second
column must be all the values of those variables. The third column must be
the proportion population margins that these values of these variables
should be weighted to. The levels must match the levels in the data. See
the example targets object for an illustration.}

\item{na_warn}{Logical. Do you want to be warned when cases are removed
due to missing values?}

\item{...}{Arguments passed to survey::rake}
}
\description{
Get rake weights, given data.frame objects representing the actual data
as well as the population marginal targets that one wishes to weight to.
Raking will be done marginally across variables specified. 
Note that missing values are deleted listwise. Any cases having missing
values in any of the variable listed in the targets object will be removed.
}
\examples{
set.seed(1839)
n <- 200
gender <- c(rep("f", n * .3), rep("m", n * .7))
party <- c(rep("r", n * .4), rep("d", n * .6))
data <- data.frame(
  gender = gender[order(runif(n))],
  party = party[order(runif(n))]
)
targets <- data.frame(
variable = c("gender", "gender", "party", "party"),
  value = c("f", "m", "d", "r"),
  pop_margin = .5
)
get_weights(data, targets)
}
