% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_map.R
\name{ks_map}
\alias{ks_map}
\title{Quickly-Made County-Level Choropleth Map for the State of Kansas}
\usage{
ks_map(data, key, variable, label = "", lowcolor = "#56B1F7",
  highcolor = "#132B43", title = "")
}
\arguments{
\item{data}{The data frame containing the variables you want to map}

\item{key}{Either “fips” or “county.” The function will throw an error if the key is not set to one of these two. It should be specified as a character string.}

\item{variable}{The variable you want to use as the fill for each county. This should be supplied as a character string.}

\item{label}{The legend title you wish to have; the label for your variable. This defaults to being blank.}

\item{lowcolor}{The color you wish to represent the minimum value of variable. This takes a hex value (or a string known by R) as a character string. It defaults to the ggplot2 default.}

\item{highcolor}{The color you wish to represent the maximum value of variable. This takes a hex value (or a string known by R) as a character string. It defaults to the ggplot2 default.}

\item{title}{The title that will go above your map. This defaults to being blank.}
}
\description{
This function requires a certain data structure: A variable needs to be labeled “fips” (with five-digit county FIPS codes) or labeled “county” (with county names, in all lowercase); additionally, it requires some numeric vector whose values you will use to fill in the choropleth map.
}
