% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_ss.R
\name{mlm_ss}
\alias{mlm_ss}
\title{Generating Simple Slopes for Two-Way Interactions in Two-Level Multilevel Models}
\usage{
mlm_ss(model, x2_1, x2_2 = NA, x2_3 = NA)
}
\arguments{
\item{model}{An lmer object created by the lme4 package. The formula of fixed effects you specify must follow a specific
pattern: "dv ~ focal*moderator + (...)", where "dv" is the dependent variable, "focal" is the focal predictor, and "moderator"
is the moderating variable. Currently, this methodology does not handle additional covariates.}

\item{x2_1}{A numeric value of the moderator for which you would like the simple slope of the focal variable.}

\item{x2_2}{A numeric value of the moderator for which you would like the simple slope of the focal variable. Optional.}

\item{x2_3}{A numeric value of the moderator for which you would like the simple slope of the focal variable. Optional.}
}
\description{
This function generates a list of two data frames: first, significance tests for simple slopes in
multilevel model interactions; second, intercepts and slopes for plotting these simple slopes.
The calculations behind this function are identical to the online calculator found at:
http://www.quantpsy.org/interact/mlr2.htm. The associated paper with these calculations can
be found at: http://www.quantpsy.org/pubs/preacher_curran_bauer_2006.pdf. Currently, this function
has been verified to work for two cases ONLY: first, "Case 1" (as specified in the Preacher et al., 2006 paper),
second, "Case 3", when the moderator is at Level-2 and the focual predictor at Level-1. Make sure your
categorical variables are numeric.
}
