% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_table.R
\name{t_table}
\alias{t_table}
\title{Produce t-test summary with means, standard deviations, statistics, Cohen's d, and confidence intervals}
\usage{
t_table(data, dvs, iv, varequal = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{dvs}{A variable name (or list of variable names) that are to be used as dependent variables.}

\item{iv}{A variable name for the condition. The corresponding variable in the data.frame should be a factor with two levels.}

\item{varequal}{A boolean value for if variances should be assumed to be equal across conditions or not.}
}
\description{
This function generates a table that includes variable names, means and standard deviations for each condition,
t-value and p-value, degrees of freedom, and Cohen's d with confidence intervals.
}
\examples{
data <- data.frame(
  a = c(rnorm(50, 0, 2), rnorm(50, 1, 2)),
  b = c(rnorm(50, 0, 2), rnorm(50, 1, 2)),
  c = c(rnorm(50, 0, 2), rnorm(50, 1, 2)),
  d = factor(c(rep("cond_0", 50), rep("cond_1", 50)))
)
t_table(data = data, dvs = c("a", "b", "c"), iv = "d")
}
